Preserve editor Vitest projects in root tests
